+++ src/includes/classes/App.php
@@ -183,7 +183,7 @@ class App extends Classes\Core\Base\Core
      *
      * @var string Version.
      */
-    const VERSION = '170408.64876'; //v//
+    const VERSION = '170412.21827'; //v//
 
     /**
      * Constructor.