+++ tests/local/clips.php
@@ -8,10 +8,10 @@ require_once dirname(__FILE__, 2).'/includes/local.php';
 
 /* ------------------------------------------------------------------------------------------------------------------ */
 
-echo c::clip('This is a really, really long sentence with a lot of characters. More than 80, that is.')."\n";
-echo c::clip('This is a really, really long sentence with a lot of characters. More than 80, that is.', 80, false, ' <a href="#">more</a>')."\n";
-echo c::clip('This is a really, really long sentence with a lot of characters. More than 80, that is.', 10, false, ' <a href="#">more</a>')."\n";
+echo c::clip('This is a really, really long sentence with a lot of characters. More than 40, that is.')."\n";
+echo c::clip('This is a really, really long sentence with a lot of characters. More than 40, that is.', 40, false, ' <a href="#">more</a>')."\n";
+echo c::clip('This is a really, really long sentence with a lot of characters. More than 40, that is.', 10, false, ' <a href="#">more</a>')."\n";
 echo "\n";
-echo c::midClip('This is a really, really long sentence with a lot of characters. More than 80, that is.')."\n";
-echo c::midClip('This is a really, really long sentence with a lot of characters. More than 80, that is.', 80, ' <a href="#">more</a>')."\n";
-echo c::midClip('This is a really, really long sentence with a lot of characters. More than 80, that is.', 10, ' <a href="#">more</a>')."\n";
+echo c::midClip('This is a really, really long sentence with a lot of characters. More than 40, that is.')."\n";
+echo c::midClip('This is a really, really long sentence with a lot of characters. More than 40, that is.', 40, ' <a href="#">more</a>')."\n";
+echo c::midClip('This is a <really> &lt;em&gt;, really long sentence with a lot of characters. More than 40, that is.', 40, ' <a href="#">more</a>')."\n";